backend/requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
faiss-cpu
openai
pydantic
sqlalchemy
fastapi
uvicorn
jinja2
PyPDF2
sqlalchemy>=2.0
psycopg2-binary
